refactor(workspace): straggler sweep after markdown subpath split
Four stale boundaries left over from the four-commit cleanup:

- SerializeResult lived in src/markdown/assemble-markdown.ts and was
  exported from @epicenter/workspace/markdown, but assembleMarkdown
  returns string; the type is the materializer's KV serializer return
  shape with zero external consumers. Inlined at the one use site in
  materializer.ts as { filename: string; content: string }.

- MaterializerWriteError (const + type) was exported from materializer.ts
  but never crossed the file boundary; it is constructed in two .catch
  blocks and passed to log.warn. Dropped the export keyword on both.

- serializers.ts (plural) held one function (slugFilename) after the
  earlier exports moved to src/markdown/. Renamed to slug-filename.ts
  to match the sqlite materializer's "one focused concern per file"
  pattern (cf. fts.ts).

- playground/opensidian-e2e/markdown-utils.ts (plural) held one
  function (prepareMarkdownFiles). Renamed to prepare-markdown-files.ts.
